+++ tests/testthat/test-verbosity.R
@@ -7,3 +7,18 @@ test_that("verbosity option is validated", {
     "must be set to one of"
   )
 })
+
+test_that("verbosity option has precedence over user env (#113)", {
+  mytool <- function() {
+    deprecate_soft("1.0.0", "mytool()")
+    10 * 10
+  }
+
+  rlang::local_options(lifecycle_verbosity = "error")
+
+  expect_error(
+    exec(mytool, .env = global_env()),
+    class = "defunctError"
+  )
+})
+